+++ applications/fishing-map/src/features/app/App.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useCallback, useEffect, Suspense, useLayoutEffect } from 'react'
+import React, { lazy, useState, useCallback, useEffect, Suspense, useLayoutEffect } from 'react'
 import { useSelector } from 'react-redux'
 // import RecoilizeDebugger from 'recoilize'
 import SplitView from '@globalfishingwatch/ui-components/dist/split-view'
@@ -16,34 +16,35 @@ import menuBgImage from 'assets/images/menubg.jpg'
 import { useLocationConnect } from 'routes/routes.hook'
 import DebugMenu from 'features/debug/DebugMenu'
 import Sidebar from 'features/sidebar/Sidebar'
-import Map from 'features/map/Map'
-import Timebar from 'features/timebar/Timebar'
 import Footer from 'features/footer/Footer'
 import {
   selectCurrentWorkspaceId,
   selectWorkspaceCustomStatus,
   selectWorkspaceStatus,
 } from 'features/workspace/workspace.selectors'
-import { fetchUserThunk, selectUserData } from 'features/user/user.slice'
+import { fetchUserThunk } from 'features/user/user.slice'
 import { fetchHighlightWorkspacesThunk } from 'features/workspaces-list/workspaces-list.slice'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import useViewport, { useMapFitBounds } from 'features/map/map-viewport.hooks'
 import { selectIsAnalyzing } from 'features/analysis/analysis.selectors'
 import { isUserLogged } from 'features/user/user.selectors'
 import { DEFAULT_WORKSPACE_ID } from 'data/workspaces'
-import { HOME, WORKSPACE } from 'routes/routes'
+import { HOME, WORKSPACE, USER, WORKSPACES_LIST } from 'routes/routes'
 import { fetchWorkspaceThunk } from 'features/workspace/workspace.slice'
+import { t } from 'features/i18n/i18n'
 import { useAppDispatch } from './app.hooks'
 import { selectAnalysisQuery, selectSidebarOpen } from './app.selectors'
 import styles from './App.module.css'
+import { useAnalytics } from './analytics.hooks'
+
+const Map = lazy(() => import(/* webpackChunkName: "Map" */ 'features/map/Map'))
+const Timebar = lazy(() => import(/* webpackChunkName: "Timebar" */ 'features/timebar/Timebar'))
 
 declare global {
   interface Window {
     gtag: any
   }
 }
-
-const GOOGLE_UNIVERSAL_ANALYTICS_ID = process.env.REACT_APP_GOOGLE_UNIVERSAL_ANALYTICS_ID
 
 const Main = () => {
   const workspaceLocation = useSelector(isWorkspaceLocation)
@@ -58,8 +59,8 @@ const Main = () => {
 }
 
 function App(): React.ReactElement {
+  useAnalytics()
   const dispatch = useAppDispatch()
-  const userData = useSelector(selectUserData)
   const sidebarOpen = useSelector(selectSidebarOpen)
   const { dispatchQueryParams } = useLocationConnect()
   const [menuOpen, setMenuOpen] = useState(false)
@@ -125,38 +126,6 @@ function App(): React.ReactElement {
   const { debugActive, dispatchToggleDebugMenu } = useDebugMenu()
 
   useEffect(() => {
-    if (userData && GOOGLE_UNIVERSAL_ANALYTICS_ID && window.gtag) {
-      window.gtag('config', GOOGLE_UNIVERSAL_ANALYTICS_ID, {
-        user_id: userData.id,
-        custom_map: {
-          dimension1: 'userId',
-          dimension3: 'userGroup',
-          dimension4: 'userOrgType',
-          dimension5: 'userOrganization',
-          dimension6: 'userCountry',
-          dimension7: 'userLanguage',
-        },
-      })
-      window.gtag('set', 'user_properties', {
-        user_id: userData.id,
-        user_group: userData.groups,
-        user_org_type: userData.organizationType,
-        user_organization: userData.organization,
-        user_country: userData.country,
-        user_language: userData.language,
-      })
-      window.gtag('event', 'login', {
-        userId: userData.id,
-        userGroup: userData.groups,
-        userOrgType: userData.organizationType,
-        userOrganization: userData.organization,
-        userCountry: userData.country,
-        userLanguage: userData.language,
-      })
-    }
-  }, [userData])
-
-  useEffect(() => {
     dispatch(fetchUserThunk())
   }, [dispatch])
 
@@ -172,6 +141,13 @@ function App(): React.ReactElement {
     setMenuOpen(true)
   }, [])
 
+  const getSidebarName = useCallback(() => {
+    if (locationType === USER) return t('user.title', 'User')
+    if (locationType === WORKSPACES_LIST) return t('workspace.title_plural', 'Workspaces')
+    if (isAnalysing) return t('analysis.title', 'Analysis')
+    return t('common.layerList', 'Layer list')
+  }, [isAnalysing, locationType])
+
   return (
     /* Value as null as there is no needed to set a default value but Typescript complains */
     <MapContext.Provider value={null as any}>
@@ -183,6 +159,8 @@ function App(): React.ReactElement {
           aside={<Sidebar onMenuClick={onMenuClick} />}
           main={<Main />}
           asideWidth={narrowSidebar ? '37rem' : '50%'}
+          showAsideLabel={getSidebarName()}
+          showMainLabel={t('common.map', 'Map')}
           className="split-container"
         />
       </Suspense>