+++ packages/ui-components/src/choice/index.ts
@@ -3,4 +3,5 @@ export { default } from './Choice'
 export interface ChoiceOption {
   id: string
   title: string
+  disabled?: boolean
 }