+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.7.82
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/react-hooks@6.3.0
+
+## 0.7.81
+
+### Patch Changes
+
+- Updated dependencies [6b374674]
+  - @globalfishingwatch/layer-composer@6.1.3
+  - @globalfishingwatch/timebar@1.5.3
+  - @globalfishingwatch/react-hooks@6.2.3
+
+## 0.7.80
+
+### Patch Changes
+
+- Updated dependencies [d43610ba]
+  - @globalfishingwatch/layer-composer@6.1.2
+  - @globalfishingwatch/react-hooks@6.2.2
+
 ## 0.7.79
 
 ### Patch Changes