+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/layer-composer
 
+## 6.1.3
+
+### Patch Changes
+
+- 6b374674: Fishing map/minor fixes
+
+## 6.1.2
+
+### Patch Changes
+
+- d43610ba: Breaks fixes
+
 ## 6.1.1
 
 ### Patch Changes