+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @globalfishingwatch/react-hooks
 
+## 6.3.0
+
+### Minor Changes
+
+- new use-small-screen hook
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/ui-components@3.1.4
+
+## 6.2.3
+
+### Patch Changes
+
+- Updated dependencies [6b374674]
+  - @globalfishingwatch/dataviews-client@6.5.2
+  - @globalfishingwatch/layer-composer@6.1.3
+  - @globalfishingwatch/ui-components@3.1.3
+
+## 6.2.2
+
+### Patch Changes
+
+- Updated dependencies [d43610ba]
+  - @globalfishingwatch/layer-composer@6.1.2
+  - @globalfishingwatch/ui-components@3.1.2
+
 ## 6.2.1
 
 ### Patch Changes