+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @globalfishingwatch/ui-components
 
+## 3.1.4
+
+### Patch Changes
+
+- support disabled choice option
+
+## 3.1.3
+
+### Patch Changes
+
+- 6b374674: Fishing map/minor fixes
+- Updated dependencies [6b374674]
+  - @globalfishingwatch/layer-composer@6.1.3
+
+## 3.1.2
+
+### Patch Changes
+
+- Updated dependencies [d43610ba]
+  - @globalfishingwatch/layer-composer@6.1.2
+
 ## 3.1.1
 
 ### Patch Changes